CIV-7617: Set Up Common Method
